==> R/11_figures.R <==
################################################# FIGURES ##############################################################################

source("R/01_helper_functions.R")

load("data/modeling_data.Rdata")

################################### AIRBNB ANALYSIS ########################################

cities_table %>% 
  dplyr::select(c("city", "active_listings_avg")) %>% 
  arrange(desc(active_listings_avg))

cities_table %>% 
  dplyr::select(c("region", "city", "active_listings_avg")) %>% 
  group_by(region) %>% 
  summarize(active_listings_avg_region = sum(active_listings_avg, na.rm = TRUE)) %>% 
  arrange(desc(active_listings_avg_region))

cities_table %>% 
  mutate(active_listings_per_household = active_listings_avg/households) %>% 
  dplyr::select(c("city", "active_listings_per_household")) %>% 
  arrange(desc(active_listings_per_household))

cities_table %>% 
  dplyr::select(c("region", "city", "active_listings_avg", "households")) %>% 
  group_by(region) %>% 
  summarize(active_listings_avg_region = sum(active_listings_avg, na.rm = TRUE), 
            households_region = sum(households, na.rm = TRUE)) %>% 
  mutate(active_listings_per_household = active_listings_avg_region/households_region) %>% 
  dplyr::select(c("region", "active_listings_per_household")) %>% 
  arrange(desc(active_listings_per_household))

cities_table %>% 
  dplyr::select(c("city", "revenue_LTM")) %>% 
  arrange(desc(revenue_LTM))

cities_table %>% 
  st_drop_geometry() %>% 
  dplyr::select(c("region", "city", "revenue_LTM")) %>% 
  group_by(region) %>% 
  summarize(revenue_LTM_region = sum(as.numeric(revenue_LTM, na.rm = TRUE))) %>% 
  arrange(desc(revenue_LTM_region))

cities_table %>% 
  dplyr::select(c("city", "revenue_LTM_per_listing")) %>% 
  arrange(desc(revenue_LTM_per_listing))

cities_table %>% 
  st_drop_geometry() %>% 
  dplyr::select(c("region", "city", "revenue_LTM", "active_listings_avg")) %>% 
  group_by(region) %>% 
  summarize(revenue_LTM_per_listing_region = sum(as.numeric(revenue_LTM, na.rm = TRUE))/sum(active_listings_avg, na.rm = TRUE)) %>% 
  arrange(desc(revenue_LTM_per_listing_region))

cities_table %>% 
  #filter(active_listings_avg > 5000) %>% 
  dplyr::select(c("city", "active_listings_yoy")) %>% 
  arrange(desc(active_listings_yoy))

cities_table %>% 
  st_drop_geometry() %>% 
  dplyr::select(c("region", "city", "active_listings_prev", "active_listings")) %>% 
  group_by(region) %>% 
  summarize(active_listings_yoy_region = sum(as.numeric(active_listings, na.rm = TRUE))/sum(active_listings_prev, na.rm = TRUE)) %>% 
  arrange(desc(active_listings_yoy_region))

cities_table %>% 
  dplyr::select(c("city", "housing_loss")) %>% 
  arrange(desc(housing_loss))

cities_table %>% 
  st_drop_geometry() %>% 
  dplyr::select(c("region", "city", "housing_loss")) %>% 
  group_by(region) %>% 
  summarize(housing_loss_region = sum(housing_loss, na.rm = TRUE))%>% 
  arrange(desc(housing_loss_region))

cities_table %>% 
  dplyr::select(c("city", "housing_loss_pct_households")) %>% 
  arrange(desc(housing_loss_pct_households))

cities_table %>% 
  st_drop_geometry() %>% 
  dplyr::select(c("region", "city", "housing_loss", "households")) %>% 
  group_by(region) %>% 
  summarize(housing_loss_pct_households_region = sum(housing_loss, na.rm = TRUE)/sum(households, na.rm = TRUE))%>% 
  arrange(desc(housing_loss_pct_households_region))

airbnb_cities <- 
  cities_table %>% 
  st_drop_geometry() %>% 
  filter(city %in% c("Anaheim", "Atlanta", "Austin", "Bakersfield",
                    "Boston", "Chesapeake", "Chicago", "El Paso",
                    "Fort Worth", "Garland", "Honolulu", "Houston",
                    "Jersey City", "Laredo", "Los Angeles", "Miami",
                    "Milwaukee", "Nashville", "New Orleans", "New York",
                    "Norfolk", "Oxnard", "Portland", "Saint Petersburg", 
                    "San Diego", "San Francisco", "Santa Clarita", "Scottsdale", 
                    "Seattle", "Virginia Beach", "Washington")) %>% 
  mutate(active_listings_per_household = active_listings_avg/households) %>% 
  dplyr::select(c(city, active_listings_avg, active_listings_per_household, 
                  revenue_LTM, revenue_LTM_per_listing, active_listings_yoy,
                  housing_loss, housing_loss_pct_households)) 

write.csv(airbnb_cities, file = "airbnb_cities.csv")


# Visualizing STR activity

colour_scale <- 
  tibble(
  "3 - 3" = "#3F2949", # high var1, high var2
  "2 - 3" = "#435786",
  "1 - 3" = "#4885C1", # low var1, high var2
  "3 - 2" = "#77324C",
  "2 - 2" = "#806A8A", # medium var1, medium var2
  "1 - 2" = "#89A1C8",
  "3 - 1" = "#AE3A4E", # high var1, low var2
  "2 - 1" = "#BC7C8F",
  "1 - 1" = "#CABED0")

distribution_listings <- 
  cities_table %>% 
  mutate(country = "") %>% 
  ggplot(aes(x = country, y = active_listings_avg)) + 
  geom_violin(fill = "#3F2949",
              colour = "#CABED0",
              lwd = 0.25,
              scale = "count", 
              adjust = 1,
              trim = TRUE,
              draw_quantiles = c(0.25, 0.75),
              linetype = "dashed") +
  geom_violin(fill = "transparent",
              colour = "#CABED0",
              lwd = 1.25,
              scale = "count", 
              adjust = 1,
              trim = TRUE,
              draw_quantiles = c(0.5)) +
  xlab("") +
  ylab("Number of listings\n") +
 # ggtitle("STR market size by active listings in 2019 per city in the United States") +
  scale_y_continuous(labels = comma) +
  theme_minimal() +
  theme(text = element_text(family = "Helvetica Light", size = 14),
        plot.title = element_text(hjust = 0.5), 
        panel.grid.major.x = element_blank(),
        panel.grid.minor.x = element_blank())

distribution_housingloss <- 
  cities_table %>% 
  mutate(country = "") %>% 
  ggplot(aes(x = country, y = housing_loss)) + 
  geom_violin(fill = "#3F2949",
              colour = "#CABED0",
              alpha = 1,
              lwd = 0.25,
              scale = "count", 
              adjust = 1,
              trim = TRUE,
              draw_quantiles = c(0.25, 0.75),
              linetype = "dashed") +
  geom_violin(fill = "transparent",
              colour = "#CABED0",
              lwd = 1.25,
              scale = "count", 
              adjust = 1,
              trim = TRUE,
              draw_quantiles = c(0.5)) +
  xlab("") +
  ylab("Housing lost to STRs\n") +
 # ggtitle("STR-induced housing loss in 2019 per city in the United States") +
  scale_y_continuous(labels = comma) +
  theme_minimal() +
  theme(text = element_text(family = "Helvetica Light", size = 14),
        plot.title = element_text(hjust = 0.5), 
        panel.grid.major.x = element_blank(),
        panel.grid.minor.x = element_blank())

distribution_revenue <- 
  cities_table %>% 
  mutate(country = "") %>% 
  ggplot(aes(x = country, y = revenue_LTM)) + 
  geom_violin(fill = "#3F2949",
              colour = "#CABED0",
              lwd = 0.25,
              scale = "count", 
              adjust = 1,
              trim = TRUE,
              draw_quantiles = c(0.25, 0.75),
              linetype = "dashed") +
  geom_violin(fill = "transparent",
              colour = "#CABED0",
              lwd = 1.25,
              scale = "count", 
              adjust = 1,
              trim = TRUE,
              draw_quantiles = c(0.5)) +
  xlab("") +
  ylab("Revenue\n") +
 # ggtitle("Revenues generated from STRs in 2019 by city in the United States") +
  scale_y_continuous(labels = paste0("$", c(0, 250, 500, 750), "M"),
                     breaks = 10^6 * c(0,250,500,750)) +
  theme_minimal() +
  theme(text = element_text(family = "Helvetica Light", size = 14),
        plot.title = element_text(hjust = 0.5), 
        panel.grid.major.x = element_blank(),
        panel.grid.minor.x = element_blank())





tiff("STR_violin.tiff", units = "in", width = 20, height = 10, res = 300)


  distribution_listings + distribution_revenue + distribution_housingloss + 
  plot_annotation(title = "Distribution of active listings, revenue generated, and housing lost to STRs by city in the United States in 2019\n",
                  theme = theme (text = element_text(family = "Helvetica Light",
                                                     size = 14),
                                 plot.title = element_text(hjust = 0.5)),
                  tag_levels = "A")
  
dev.off()

states_sf <- get_urbn_map("states", sf = TRUE)

counties_sf <- 
  get_urbn_map("counties", sf = TRUE)

cities_hawaii_alaska <- 
  cities_table %>% 
  filter(city == "Anchorage" |
           city == "Honolulu") %>% 
  mutate(county_name = 
           ifelse(city == "Anchorage",
                  "Anchorage Municipality",
                  "Honolulu County"))

cities_hawaii_alaska <- 
  cities_hawaii_alaska %>% 
  st_drop_geometry() %>% 
  left_join(counties_sf) %>%
  st_as_sf() 


tiff("STR_map.tiff", units = "in", width = 20, height = 10, res = 300)

cities_table %>%
  filter(city != "Anchorage" &
           city != "Honolulu") %>% 
  ggplot() +
  geom_sf(data = states_sf, 
          colour = alpha("#3F2949", 0.75),
          fill = "transparent") +
  geom_sf(data = cities_table %>%
                    filter(city != "Anchorage" &
                          city != "Honolulu") %>%
            st_cast("POINT"),
          aes(size = active_listings, colour = region),
          alpha = 0.75) +
   geom_sf(data = cities_hawaii_alaska %>% 
             st_centroid(),
           aes(size = active_listings, colour = region),
           alpha = 0.75) +
  coord_sf(datum = NA) +
  theme_minimal() +
  scale_color_manual(values = c("Midwest" = "#435786", 
                                "Northeast" = "#77324C",
                                "South" = "#89A1C8", 
                                "West" = "#BC7C8F")) +
  ggtitle("Active STR listings throughout the United States") +
  xlab("") +
  ylab("") +
  labs(colour = "Region",
       size = "Active listings") +
  scale_size_continuous(labels = comma,
                        range = c(3, 10)) +
  scale_alpha(guide = "none") +
  guides(colour = guide_legend(override.aes = list(alpha = 0.75, size = 3))) + 
  guides(size=guide_legend(override.aes=list(colour="#CABED0"))) +
  theme(text = element_text(family = "Helvetica Light", size = 14),
        plot.title = element_text(hjust = 0.6), 
        panel.grid.major.x = element_blank(),
        panel.grid.minor.x = element_blank())

dev.off()
################################### MEDIA ANALYSIS ###########################################

##### NUMBER OF ARTICLES

media_table$Date <- as.Date(media_table$Date)

media_table$month_yr <- (format(as.Date(media_table$Date), "%Y-%m"))

media_table$month_yr <- as.Date(paste(media_table$month_yr, "-15", sep = ""))

# Country-wide number of articles over time

tiff("country_discourse.tiff", units = "in", width = 20, height = 10, res = 300)

media_table %>% 
  filter(Date >= "2015-01-01" &
           Date <= "2019-12-31") %>% 
  group_by(month_yr) %>% 
  count() %>% 
  ggplot(aes(x = month_yr, y = n)) +
  geom_line(color = "#3F2949", 
            lwd = 0.25, 
            show.legend = TRUE)+
  geom_smooth(method = "gam",
              color = "#3F2949", 
              fill = "#CABED0", 
              lwd = 2,
              show.legend = TRUE) +
  xlab("\nDate") +
  ylab("Number of news articles\n") +
  ggtitle("STR discourse throughout the United States over time\n") +
  scale_y_continuous(labels = comma) +
  theme_minimal() +
  theme(text = element_text(family = "Helvetica Light", size = 14),
        plot.title = element_text(hjust = 0.5), 
        panel.grid.major.x = element_blank(),
        panel.grid.minor.x = element_blank())

dev.off()

tiff("discourse_map.tiff", units = "in", width = 20, height = 10, res = 300)

  cities_table %>%
  filter(city != "Anchorage" &
           city != "Honolulu") %>% 
  ggplot() +
  geom_sf(data = states_sf, 
          colour = alpha("#3F2949", 0.75),
          fill = "transparent") +
  geom_sf(data = cities_table %>%
            filter(city != "Anchorage" &
                     city != "Honolulu") %>%
            st_cast("POINT"),
          aes(size = n, colour = region),
          alpha = 0.75) +
  geom_sf(data = cities_hawaii_alaska %>% 
            st_centroid(),
          aes(size = n, colour = region),
          alpha = 0.75) +
  coord_sf(datum = NA) +
  theme_minimal() +
  scale_color_manual(values = c("Midwest" = "#435786", 
                                "Northeast" = "#77324C",
                                "South" = "#89A1C8", 
                                "West" = "#BC7C8F")) +
    ggtitle("The distribution of STR discourse by city and region") +
  xlab("") +
  ylab("") +
  labs(colour = "Region",
       size = "Number of news articles") +
  scale_size_continuous(labels = comma,
                        range = c(3, 10)) +
  scale_alpha(guide = "none") +
  guides(colour = guide_legend(override.aes = list(alpha = 0.75, size = 3))) + 
  guides(size = guide_legend(override.aes = list(colour = "#CABED0"))) +
  theme(text = element_text(family = "Helvetica Light", size = 14),
        plot.title = element_text(hjust = 0.5), 
        panel.grid.major.x = element_blank(),
        panel.grid.minor.x = element_blank()) 

dev.off()

# Regional number of articles over time

tiff("regional_discourse.tiff", units = "in", width = 20, height = 10, res = 300)

  media_table %>% 
  filter(Date >= "2015-01-01" &
           Date <= "2019-12-31") %>% 
  group_by(region, month_yr) %>% 
  count() %>%
  ggplot(aes(x = month_yr, y = n, color = region)) +
  geom_line(lwd = 0.25, 
            show.legend = TRUE)+
  geom_smooth(aes(fill = region),
              method = "gam",
              lwd = 1,
              show.legend = TRUE,
              se = TRUE,
              alpha = 0.25) +
  xlab("\nDate") +
  ylab("\nNumber of news articles\n") +
    ggtitle("\nSTR discourse by region throughout the United States over time") +
  scale_y_continuous(breaks = c(0, 250, 750, 1000), limits = c(0, 1000), labels = comma) +
  scale_color_manual(name = "Region",
                     values = c("Midwest" = "#435786", 
                                "Northeast" = "#77324C",
                                "South" = "#89A1C8", 
                                "West" = "#BC7C8F")) +
  scale_fill_manual(name = "Region",
                    values = c("Midwest" = "#435786", 
                                "Northeast" = "#77324C",
                                "South" = "#89A1C8", 
                                "West" = "#BC7C8F")) +
  theme_minimal() +
  theme(text = element_text(family = "Helvetica Light", size = 14),
        plot.title = element_text(hjust = 0.5), 
        panel.grid.major.x = element_blank(),
        panel.grid.minor.x = element_blank(),
        panel.grid.minor.y = element_blank(),
        plot.margin = margin(t = 0, unit = "cm"))

dev.off()

discourse_map + discourse_region + plot_layout(ncol = 1,
                                               widths = c(1, 1),
                                               heights = c(2.5, 1)) +
  plot_annotation(title = "STR discourse throughout the United States\n",
                  theme = theme (text = element_text(family = "Helvetica Light",
                                                     size = 18),
                                 plot.title = element_text(hjust = 0.5)))

# Number of articles by city

cities_media <-
  media_table %>%
  group_by(city) %>%
  count() %>%
  arrange(desc(n))
# 
# cities <- 
#   cities_media[1:4,1] %>% 
#   do.call(paste, .)
# 
# data <- 
#   media_table %>% 
#   filter(Date >= "2015-01-01" &
#            Date <= "2019-12-31") %>% 
#   filter(city %in% cities) %>% 
#   group_by(city, month_yr) %>% 
#   count() 
# 
# data$city <- factor(data$city, levels = c("New York",
#                                           "San Francisco",
#                                           "Washington",
#                                           "Los Angeles"))
#   
# data %>%
#   ggplot(aes(x = month_yr, y = n)) +
#   geom_area(fill = "#CABED0", alpha = 0.5) +
#   stat_smooth(
#     geom = 'area', method = 'loess', span = 1/5,
#     alpha = 0.5, fill = "#3F2949") +
#   facet_grid(vars(city)) +
#   xlab("\nDate") +
#   ylab("Number of news articles\n") +
#   ggtitle("STR discourse by city throughout the United States over time\n") +
#   scale_y_continuous(labels = comma) +
#   theme_minimal() +
#   theme(text = element_text(family = "Helvetica Light", size = 14),
#         plot.title = element_text(hjust = 0.5), 
#         panel.grid.major.x = element_blank(),
#         panel.grid.minor.x = element_blank(), 
#         panel.grid.minor.y = element_blank())
# 
# cities <- 
#   cities_media[5:13,1] %>% 
#   do.call(paste, .)
# 
# media_table %>% 
#   filter(Date >= "2015-01-01" &
#            Date <= "2019-12-31") %>% 
#   filter(city %in% cities) %>% 
#   group_by(city) %>% 
#   count() %>% 
#   arrange(n)
# 
# data <- 
#   media_table %>% 
#   filter(Date >= "2015-01-01" &
#            Date <= "2019-12-31") %>% 
#   filter(city %in% cities) %>% 
#   group_by(city, month_yr) %>% 
#   count() 
# 
# data$city <- factor(data$city, levels = c("Chicago",
#                                           "Boston",
#                                           "Miami", 
#                                           "Seattle",
#                                           "Austin", 
#                                           "Las Vegas",
#                                           "Houston",
#                                           "Philadelphia",
#                                           "San Diego"))
# data %>%
#   ggplot(aes(x = month_yr, y = n)) +
#   geom_area(fill = "#CABED0", alpha = 0.5) +
#   stat_smooth(
#     geom = 'area', method = 'loess', span = 1/5,
#     alpha = 0.5, fill = "#3F2949") +
#   facet_grid(vars(city)) +
#   xlab("\nDate") +
#   ylab("Number of news articles\n") +
#   ggtitle("STR discourse by city throughout the United States over time\n") +
#   scale_y_continuous(breaks = c(0, 50, 100), labels = comma) +
#   theme_minimal() +
#   theme(text = element_text(family = "Helvetica Light", size = 14),
#         plot.title = element_text(hjust = 0.5), 
#         panel.grid.major.x = element_blank(),
#         panel.grid.minor.x = element_blank(), 
#         panel.grid.minor.y = element_blank())


cities <- 
  cities_media[1:9,1] %>% 
  do.call(paste, .)

city_media_absolute <- 
  media_table %>% 
  filter(Date >= "2015-01-01" &
           Date <= "2019-12-31") %>% 
  filter(city %in% cities) %>% 
  group_by(city, month_yr) %>% 
  count() %>%
  ggplot(aes(x = month_yr, y = n, color = city)) +
  geom_smooth(
              method = "gam",
              lwd = 1,
              se = FALSE,
              alpha = 0.25) +
  xlab(NULL) +
  ylab("\nNumber of news articles\n") +
  # ggtitle("\nSTR discourse by city over time") +
  scale_y_continuous(labels = comma) +
  scale_color_manual(name = "City",
                     values = c("Austin" = "#3F2949",
                                "Boston" = "#435786",
                                "Chicago" = "#4885C1",
                                "Los Angeles" = "#77324C",
                                "Miami" = "#806A8A",
                                "New York" = "#89A1C8",
                                "San Francisco" = "#AE3A4E",
                                "Seattle" = "#BC7C8F",
                                "Washington" = "#CABED0")) +
  theme_minimal() +
  theme(text = element_text(family = "Helvetica Light", size = 14),
        plot.title = element_text(hjust = 0.5), 
        panel.grid.major.x = element_blank(),
        panel.grid.minor.x = element_blank(),
        panel.grid.minor.y = element_blank(),
        axis.text.x = element_blank())

city_media_pop<- 
  cities_table %>% 
  filter(city %in% cities) %>% 
  dplyr::select(city, population, active_listings) %>% 
  st_drop_geometry() %>% 
  right_join(
    media_table %>% 
      filter(Date >= "2015-01-01" &
               Date <= "2019-12-31") %>% 
      filter(city %in% cities) %>% 
      group_by(city, month_yr) %>% 
      count()) %>% 
  mutate(articles_per_capita = n/population,
         articles_per_listing = n/active_listings) %>% 
  ggplot(aes(x = month_yr, y = articles_per_capita, color = city)) +
  geom_smooth(
    method = "gam",
    lwd = 1,
    se = FALSE,
    alpha = 0.25) +
  xlab(NULL) +
  ylab("\nNumber of news articles per capita\n") +
  # ggtitle("\nSTR discourse per capita by city over time") +
  scale_y_continuous(labels = scales::number_format(accuracy = 0.00001)) +
  scale_color_manual(name = "City",
                     values = c("Austin" = "#3F2949",
                                "Boston" = "#435786",
                                "Chicago" = "#4885C1",
                                "Los Angeles" = "#77324C",
                                "Miami" = "#806A8A",
                                "New York" = "#89A1C8",
                                "San Francisco" = "#AE3A4E",
                                "Seattle" = "#BC7C8F",
                                "Washington" = "#CABED0")) +
  theme_minimal() +
  theme(text = element_text(family = "Helvetica Light", size = 14),
        plot.title = element_text(hjust = 0.5), 
        panel.grid.major.x = element_blank(),
        panel.grid.minor.x = element_blank(),
        panel.grid.minor.y = element_blank(),
        axis.text.x = element_blank())

city_media_listing <- 
  cities_table %>% 
  filter(city %in% cities) %>% 
  dplyr::select(city, population, active_listings) %>% 
  st_drop_geometry() %>% 
  right_join(
    media_table %>% 
      filter(Date >= "2015-01-01" &
               Date <= "2019-12-31") %>% 
      filter(city %in% cities) %>% 
      group_by(city, month_yr) %>% 
      count()) %>% 
  mutate(articles_per_capita = n/population,
         articles_per_listing = n/active_listings) %>% 
  ggplot(aes(x = month_yr, y = articles_per_listing, color = city)) +
  geom_smooth(
    method = "gam",
    lwd = 1,
    se = FALSE,
    alpha = 0.25) +
  xlab("\nDate") +
  ylab("\nNumber of news articles per listing\n") +
  # ggtitle("\nSTR discourse per listing by city over time") +
  scale_y_continuous(labels = scales::number_format(accuracy = 0.01)) +
  scale_color_manual(name = "City",
                     values = c("Austin" = "#3F2949",
                                "Boston" = "#435786",
                                "Chicago" = "#4885C1",
                                "Los Angeles" = "#77324C",
                                "Miami" = "#806A8A",
                                "New York" = "#89A1C8",
                                "San Francisco" = "#AE3A4E",
                                "Seattle" = "#BC7C8F",
                                "Washington" = "#CABED0")) +
  theme_minimal() +
  theme(text = element_text(family = "Helvetica Light", size = 14),
        plot.title = element_text(hjust = 0.5), 
        panel.grid.major.x = element_blank(),
        panel.grid.minor.x = element_blank(),
        panel.grid.minor.y = element_blank())
  

tiff("cities_discourse.tiff", units = "in", width = 20, height = 20, res = 300)

city_media_absolute + plot_spacer() +
  city_media_pop + plot_spacer() +
  city_media_listing +
  plot_layout(ncol = 1, guides = "collect", 
              heights = c(1, 0.25, 1, 0.25, 1)) +
  plot_annotation(title = "STR discourse by city over time\n",
                  theme = theme (text = element_text(family = "Helvetica Light",
                                                     size = 14),
                                 plot.title = element_text(hjust = 0.5)),
                  tag_levels = "A")

dev.off()

##### SENTIMENT


media_table %>% 
  mutate(sentiment_round = round(media_table$sentiment, digits = 2)) %>% 
  filter(Date >= "2015-01-01" &
         Date <= "2019-12-31") %>% 
  group_by(sentiment_round) %>% 
  count() %>% 
  ggplot(aes(x = n, y = sentiment_round)) + 
  geom_point()

# Country-wide sentiment of articles over time

tiff("country_sentiment.tiff", units = "in", width = 20, height = 10, res = 300)

media_table %>% 
  filter(Date >= "2015-01-01" &
           Date <= "2019-12-31") %>% 
  # group_by(month_yr) %>% 
  # summarize(sentiment_mean = mean(sentiment, na.rm = TRUE)) %>% 
  ggplot(aes(Date, sentiment)) + 
  # geom_line(color = "#3F2949", 
  #           lwd = 0.25) +
  geom_smooth(color = "#3F2949", 
              fill = "#CABED0", 
              lwd = 2) +
  xlab("\nDate") +
  ylab("Sentiment\n") +
  ggtitle("The sentiment of STR discourse throughout the United States over time\n") +
  scale_y_continuous(breaks = 0, labels = comma) +
  theme_minimal() +
  theme(text = element_text(family = "Helvetica Light", size = 14),
        plot.title = element_text(hjust = 0.5), 
        panel.grid.major.x = element_blank(),
        panel.grid.minor.x = element_blank())

dev.off()

# Regional sentiment of articles over time

tiff("regional_sentiment.tiff", units = "in", width = 20, height = 10, res = 300)

media_table %>% 
  filter(Date >= "2015-01-01" &
           Date <= "2019-12-31") %>% 
ggplot(aes(Date, sentiment, color = region, fill = region)) +
  #geom_point() +
  geom_smooth(lwd = 1,
              alpha = 0.25) +
  # geom_smooth(data = media_table %>%
  #               filter(Date >= "2015-01-01" &
  #                        Date <= "2019-12-31"),
  #             aes(color = "#3F2949", fill = "#3F2949", lwd = 1),
  #             show.legend = TRUE) +
  xlab("\nDate") +
  ylab("Sentiment\n") +
  ggtitle("The sentiment of STR discourse by region throughout the United States over time\n") +
  scale_y_continuous(breaks = 0, labels = comma) +
  scale_color_manual(name = "Region",
                     values = c("Midwest" = "#435786", 
                                "Northeast" = "#77324C",
                                "South" = "#89A1C8", 
                                "West" = "#BC7C8F")) +
  scale_fill_manual(name = "Region",
                    values = c("Midwest" = "#435786", 
                               "Northeast" = "#77324C",
                               "South" = "#89A1C8", 
                               "West" = "#BC7C8F")) +
  theme_minimal() +
  theme(text = element_text(family = "Helvetica Light", size = 14),
        plot.title = element_text(hjust = 0.5), 
        panel.grid.major.x = element_blank(),
        panel.grid.minor.x = element_blank())

dev.off()

# City sentiment of articles over time

cities_media <- 
  media_table %>% 
  group_by(city) %>% 
  count() %>% 
  arrange(desc(n))

# cities <- 
#   cities_media[1:4,1] %>% 
#   do.call(paste, .)
# 
# data <- 
#   media_table %>% 
#   filter(Date >= "2015-01-01" &
#            Date <= "2019-12-31") %>% 
#   filter(city %in% cities)
# 
# data$city <- factor(data$city, levels = c("New York",
#                                           "San Francisco",
#                                           "Washington",
#                                           "Los Angeles"))
# 
# data %>% 
#   ggplot(aes(Date, sentiment)) +
#   geom_smooth(span = 0.8, 
#               color = "#3F2949", 
#               fill = "#CABED0", 
#               lwd = 1) +
#   facet_grid(vars(city)) +
#   xlab("\nDate") +
#   ylab("Sentiment\n") +
#   ggtitle("The sentiment of STR discourse by city throughout the United States over time\n") +
#   scale_y_continuous(breaks = 0, labels = comma) +
#   theme_minimal() +
#   theme(text = element_text(family = "Helvetica Light", size = 14),
#         plot.title = element_text(hjust = 0.5), 
#         panel.grid.major.x = element_blank(),
#         panel.grid.minor.x = element_blank())
# 
# cities <- 
#   cities_media[5:13,1] %>% 
#   do.call(paste, .)
# 
# data <- 
#   media_table %>% 
#   filter(Date >= "2015-01-01" &
#            Date <= "2019-12-31") %>% 
#   filter(city %in% cities)
# 
# data$city <- factor(data$city, levels = c("Chicago",
#                                           "Boston",
#                                           "Miami", 
#                                           "Seattle",
#                                           "Austin", 
#                                           "Las Vegas",
#                                           "Houston",
#                                           "Philadelphia",
#                                           "San Diego"))
# 
# data %>% 
#   ggplot(aes(Date, sentiment)) +
#   geom_smooth(span = 0.8, 
#               color = "#3F2949", 
#               fill = "#CABED0", 
#               lwd = 1) +
#   facet_grid(vars(city)) +
#   xlab("\nDate") +
#   ylab("Sentiment\n") +
#   ggtitle("The sentiment of STR discourse by city throughout the United States over time\n") +
#   scale_y_continuous(breaks = 0, labels = comma) +
#   theme_minimal() +
#   theme(text = element_text(family = "Helvetica Light", size = 14),
#         plot.title = element_text(hjust = 0.5), 
#         panel.grid.major.x = element_blank(),
#         panel.grid.minor.x = element_blank())

cities <-
   cities_media[1:9,1] %>%
   do.call(paste, .)

tiff("city_sentiment.tiff", units = "in", width = 20, height = 10, res = 300)

media_table %>% 
  filter(Date >= "2015-01-01" &
           Date <= "2019-12-31") %>% 
  filter(city %in% cities) %>% 
  ggplot(aes(Date, sentiment, color = city)) +
  geom_smooth(
    lwd = 1,
    se = FALSE,
    alpha = 0.25) +
  xlab("\nDate") +
  ylab("\nSentiment\n") +
  ggtitle("\nThe sentiment of STR discourse by city throughout the United States over time") +
  scale_y_continuous(labels = comma) +
  scale_color_manual(name = "City",
                     values = c("Austin" = "#3F2949",
                                "Boston" = "#435786",
                                "Chicago" = "#4885C1",
                                "Los Angeles" = "#77324C",
                                "Miami" = "#806A8A",
                                "New York" = "#89A1C8",
                                "San Francisco" = "#AE3A4E",
                                "Seattle" = "#BC7C8F",
                                "Washington" = "#CABED0")) +
  theme_minimal() +
  theme(text = element_text(family = "Helvetica Light", size = 14),
        plot.title = element_text(hjust = 0.5), 
        panel.grid.major.x = element_blank(),
        panel.grid.minor.x = element_blank(),
        panel.grid.minor.y = element_blank())

dev.off()

##### COMMUNITY SENTIMENT INDEX

# Country CSI over time

tiff("country_CSI.tiff", units = "in", width = 20, height = 10, res = 300)

media_table %>% 
  filter(Date >= "2015-01-01" &
           Date <= "2019-12-31") %>% 
  group_by(month_yr) %>% 
  summarize(sentiment = mean(sentiment, na.rm = TRUE),
            articles = n()) %>% 
  mutate(CSI = sentiment * articles) %>% 
  ggplot(aes(month_yr, CSI)) +
  # geom_line(color = "#3F2949", 
  #           lwd = 0.25, 
  #           show.legend = TRUE)+
  geom_smooth(method = "gam",
              color = "#3F2949", 
              fill = "#CABED0", 
              lwd = 2,
              show.legend = TRUE) +
  xlab("\nDate") +
  ylab("CSI\n") +
  ggtitle("The community sentiment index throughout the United States over time\n") +
  scale_y_continuous(breaks = pretty_breaks(), labels = number_format(accuracy = 1)) +
  theme_minimal() +
  theme(text = element_text(family = "Helvetica Light", size = 14),
        plot.title = element_text(hjust = 0.5), 
        panel.grid.major.x = element_blank(),
        panel.grid.minor.x = element_blank())

dev.off()


# Region CSI over time

tiff("regional_CSI.tiff", units = "in", width = 20, height = 10, res = 300)

media_table %>% 
  filter(Date >= "2015-01-01" &
           Date <= "2019-12-31") %>% 
  group_by(region, month_yr) %>% 
  summarize(sentiment = mean(sentiment, na.rm = TRUE),
            articles = n()) %>% 
  mutate(CRI = -1 * sentiment * articles) %>% 
  ggplot(aes(month_yr, sentiment, color = region, fill = region)) +
  geom_smooth(method = "gam",
              lwd = 1,
              alpha = 0.25) +
  xlab("\nDate") +
  ylab("CSI\n") +
  ggtitle("The community sentiment index by region throughout the United States over time\n") +
  scale_y_continuous(breaks = c(-4, -2, 0, 2, 4), labels = number_format(accuracy = 1)) +
  scale_color_manual(name = "Region",
                     values = c("Midwest" = "#435786", 
                                "Northeast" = "#77324C",
                                "South" = "#89A1C8", 
                                "West" = "#BC7C8F")) +
  scale_fill_manual(name = "Region",
                    values = c("Midwest" = "#435786", 
                               "Northeast" = "#77324C",
                               "South" = "#89A1C8", 
                               "West" = "#BC7C8F")) +
  theme_minimal() +
  theme(text = element_text(family = "Helvetica Light", size = 14),
        plot.title = element_text(hjust = 0.5), 
        panel.grid.major.x = element_blank(),
        panel.grid.minor.x = element_blank(), 
        panel.grid.minor.y = element_blank())


dev.off()




# City CSI over time

cities_media <- 
  media_table %>% 
  group_by(city) %>% 
  count() %>% 
  arrange(desc(n))

cities <- 
  cities_media[1:9,1] %>% 
  do.call(paste, .)

tiff("city_CSI.tiff", units = "in", width = 20, height = 10, res = 300)

media_table %>% 
  filter(Date >= "2015-01-01" &
           Date <= "2019-12-31") %>% 
  filter(city %in% cities) %>% 
  group_by(city, month_yr) %>% 
  summarize(sentiment = mean(sentiment, na.rm = TRUE),
            articles = n()) %>% 
  mutate(CSI = sentiment * articles) %>% 
  ggplot(aes(month_yr, CSI, color = city)) +
  geom_smooth( method = "gam",
    lwd = 1,
    se = FALSE,
    alpha = 0.25) +
  xlab("\nDate") +
  ylab("CSI\n") +
  ggtitle("The community sentiment index by city throughout the United States over time\n") +
  scale_y_continuous(labels = number_format(accuracy = 1)) +
  scale_color_manual(name = "City",
                     values = c("Austin" = "#3F2949",
                                "Boston" = "#435786",
                                "Chicago" = "#4885C1",
                                "Los Angeles" = "#77324C",
                                "Miami" = "#806A8A",
                                "New York" = "#89A1C8",
                                "San Francisco" = "#AE3A4E",
                                "Seattle" = "#BC7C8F",
                                "Washington" = "#CABED0")) +
  theme_minimal() +
  theme(text = element_text(family = "Helvetica Light", size = 14),
        plot.title = element_text(hjust = 0.5), 
        panel.grid.major.x = element_blank(),
        panel.grid.minor.x = element_blank(),
        panel.grid.minor.y = element_blank())

dev.off()

## CRI

cities_table %>% 
  mutate(CSI = -1* CRI) %>% 
  dplyr::select(c(city, CSI, population, active_listings, housing_loss_pct_households)) %>% 
  st_drop_geometry() %>% 
  arrange((CSI)) %>% view()

neighbourhoods_table %>% 
  dplyr::select(c(city, neighbourhood, CRI)) %>% 
  st_drop_geometry() %>% 
  arrange((CRI)) %>% view()

neighbourhoods_table %>% 
  dplyr::select(c(city, neighbourhood, CRI)) %>% 
  st_drop_geometry() %>% 
  arrange(desc(CRI)) %>% view()


############################################# BIVARIATE MAPPING ################################################################
# Visualizing two variables at the same time using a bivariate colour scale

# What cities to include? 3 for each variable
  # Top 15 cities with media attention
cities_media <- 
  media_table %>% 
  group_by(city) %>% 
  count() %>% 
  arrange(desc(n)) %>% 
  dplyr::select(city) %>% 
  do.call(paste, .)

# But also those with the highest instances of the variable

# HOUSING LOSS
  # New Orleans (most housing loss), Jersey City (high housing loss), San Fran (media)
  cities_table %>% 
    arrange(desc(housing_loss_pct_households)) %>% 
    dplyr::select(city) %>% 
    st_drop_geometry() %>% 
    do.call(paste, .)

# Specify data, variables, title, labels, and quantiles (optional)
  # New Orleans
  
neighbourhoods_table <- 
  neighbourhoods_table %>% 
  mutate(CSI = -1* CRI)

data <- 
  neighbourhoods_table %>%  
  filter(city == "New Orleans") %>% 
  filter(!is.na(CRI)) %>% 
  filter(housing_loss_pct_households >= 0) %>% 
  st_as_sf() %>% 
  mutate(housing_loss_pct_households_inverse = 1/housing_loss_pct_households)

water_LA <- 
  st_read("data/tiger_la_water_CENSUS_2006/tiger_la_water_CENSUS_2006.shp") %>% 
  st_as_sf() %>% 
  st_transform(102009)

streets_neworleans <-
  (getbb("New Orleans") * c(1.01, 0.99, 0.99, 1.01)) %>%
  opq(timeout = 200) %>%
  add_osm_feature(key = "highway", value = c("primary", "secondary",
                                             "tertiary", "motorway")) %>%
  osmdata_sf()

new_orleans <- 
  bivariate_mapping(data = data,
                  streets = streets_neworleans,
                  water = water_LA,
                  buffer = 5000,
                  var1 = data$CSI, 
                  var2 = data$housing_loss_pct_households_inverse, 
                  title = "New Orleans",
                  xlab = "Increasing CSI", 
                  ylab = "Decreasing Housing Loss")

  # Jersey City
data <- 
  neighbourhoods_table %>%  
  filter(city == "Jersey City") %>% 
  filter(!is.na(CRI)) %>% 
  filter(housing_loss_pct_households >= 0) %>% 
  st_as_sf() %>% 
  mutate(housing_loss_pct_households_inverse = 1/housing_loss_pct_households)

water_NJ <- 
  st_read("data/nhdwaterbody2002shp/nhdwaterbody2002.shp") %>% 
  st_as_sf() %>% 
  st_transform(102009) %>% 
  st_set_precision(1000000) %>% 
  st_make_valid() 

streets_jersey <-
  (getbb("Jersey City") * c(1.01, 0.99, 0.99, 1.01)) %>%
  opq(timeout = 200) %>%
  add_osm_feature(key = "highway", value = c("primary", "secondary",
                                             "tertiary", "motorway")) %>%
  osmdata_sf()

jersey_city <- 
  bivariate_mapping(data = data,
                  streets = streets_jersey,
                  water = water_NJ,
                  buffer = 3000,
                  var1 = data$CSI, 
                  var2 = data$housing_loss_pct_households_inverse, 
                  title = "Jersey City",
                  xlab = "Increasing CSI", 
                  ylab = "Decreasing Housing Loss")

  # San Francisco
data_hold <- 
  neighbourhoods_table %>% 
  filter(city == "San Francisco") %>% 
  filter(!is.na(CRI)) %>% 
  filter(housing_loss_pct_households >= 0)
  
data_hold <- 
  data_hold[-c(4,5),]
  
data <- 
  neighbourhoods_table %>%  
  filter(city == "San Francisco") %>% 
  filter(!is.na(CRI)) %>% 
  filter(housing_loss_pct_households >= 0) %>% 
  st_as_sf() %>% 
  mutate(housing_loss_pct_households_inverse = 1/housing_loss_pct_households) %>% 
  st_intersection(st_buffer(st_as_sfc(st_bbox(data_hold)), 7000))

rm(data_hold)

water_CA <- 
  st_read("data/Shape_CA/NHDWaterbody.shx") %>% 
  st_as_sf() %>% 
  st_transform(102009) %>% 
  st_zm() %>% 
  st_cast("MULTIPOLYGON") %>% 
  st_make_valid() 

water_CA_ocean <- 
  st_read("data/Shape_CA/NHDArea.shp") %>% 
  st_as_sf() %>% 
  st_transform(102009) %>% 
  st_zm()

water_CA <- 
  rbind(water_CA_ocean %>% dplyr::select("geometry"),
        water_CA %>% dplyr::select("geometry"))

streets_sanfran <-
  (getbb("San Francisco") * c(1.01, 0.99, 0.99, 1.01)) %>%
  opq(timeout = 200) %>%
  add_osm_feature(key = "highway", value = c("primary", "secondary",
                                             "tertiary", "motorway")) %>%
  osmdata_sf()

san_fran <- 
  bivariate_mapping(data = data,
                  streets = streets_sanfran,
                  water = water_CA,
                  buffer = 5000,
                  var1 = data$CSI, 
                  var2 = data$housing_loss_pct_households_inverse, 
                  title = "San Francisco",
                  xlab = "Increasing CSI", 
                  ylab = "Decreasing Housing Loss")

# Housing loss legend

theme_map <- function(...) {
  theme_minimal() +
    theme(
      text = element_text(family = "Helvetica Light", size = 11),
      # remove all axes
      axis.line = element_blank(),
      axis.text.x = element_blank(),
      axis.text.y = element_blank(),
      axis.ticks = element_blank(),
      # remove grid
      panel.grid.major = element_blank(),
      panel.grid.minor = element_blank(),
      # background colors
      plot.background = element_rect(fill = "white",
                                     color = NA),
      panel.background = element_rect(fill = "white",
                                      color = NA),
      legend.background = element_rect(fill = "white",
                                       color = NA),
      # borders and margins
      plot.margin = unit(c(.5, .5, .2, .5), "cm"),
      panel.border = element_blank(),
      panel.spacing = unit(c(-.1, 0.2, .2, 0.2), "cm"),
      # titles
      legend.title = element_text(size = 11),
      legend.text = element_text(size = 10, hjust = 0,
                                 color = "black"),
      plot.title = element_text(size = 18, hjust = 0.5,
                                color = "black"),
      plot.subtitle = element_text(size = 10, hjust = 0.5,
                                   color = "black",
                                   margin = margin(b = -0.1,
                                                   t = -0.1,
                                                   l = 2,
                                                   unit = "cm"),
                                   debug = F),
      # captions
      plot.caption = element_text(size = 7,
                                  hjust = .5,
                                  margin = margin(t = 0.2,
                                                  b = 0,
                                                  unit = "cm"),
                                  color = "#939184"),
      ...
    )
}

bivariate_color_scale <- tibble(
  "3 - 3" = "#3F2949", # high var1, high var2
  "2 - 3" = "#435786",
  "1 - 3" = "#4885C1", # low var1, high var2
  "3 - 2" = "#77324C",
  "2 - 2" = "#806A8A", # medium var1, medium var2
  "1 - 2" = "#89A1C8",
  "3 - 1" = "#AE3A4E", # high var1, low var2
  "2 - 1" = "#BC7C8F",
  "1 - 1" = "#CABED0" # low var1, low var2
) %>%
  gather("group", "fill")

quantiles_var1 <- 
  quantile(data$CSI, probs = seq(0, 1, length.out = 4))

quantiles_var2 <- 
  quantile(data$housing_loss_pct_households_inverse, probs = seq(0, 1, length.out = 4))

data <- 
  data %>% 
  mutate(
    var1_quantiles = cut(
      data$CSI,
      breaks = quantiles_var1,
      include.lowest = TRUE
    ),
    var2_quantiles = cut(
      data$housing_loss_pct_households_inverse,
      breaks = quantiles_var2,
      include.lowest = TRUE
    ),
    group = paste(
      as.numeric(var1_quantiles), "-",
      as.numeric(var2_quantiles)
    )
  ) %>%
  left_join(bivariate_color_scale, by = "group")

bivariate_color_scale <- 
    bivariate_color_scale %>% 
    separate(group, into = c("var1", "var2"), sep = " - ") %>%
     mutate(var1 = as.integer(var1),
            var2 = as.integer(var2))

# Generate legend
legend <- 
  ggplot() +
  geom_tile(
    data = bivariate_color_scale,
    mapping = aes(
      x = var1,
      y = var2,
      fill = fill)
  ) +
  scale_fill_identity() +
  labs(x = "Increasing CSI",
       y = "Decreasing housing loss") +
  theme_map() +
  theme(
    axis.title = element_text(size = 10)) +
  coord_fixed()

tiff("bivariate_housingloss.tiff", units = "in", width = 15, height = 15, res = 300)


((san_fran/new_orleans) |
  (jersey_city / plot_spacer() / legend / plot_spacer() 
                  + plot_layout(heights = c(3, 0.25, 0.75, 0.25)))) +
    plot_layout(ncol = 2, widths = c(1, 1)) + 
  plot_annotation(title = "The inverse relationship between housing loss and community sentiment\n",
                  theme = theme (text = element_text(family = "Helvetica Light",
                                                     size = 18),
                                 plot.title = element_text(hjust = 0.5)))


dev.off()

### ACTIVE LISTINGS
cities_active_listings <- 
  cities_table %>% 
  arrange(desc(active_listings)) %>% 
  dplyr::select(city) %>% 
  st_drop_geometry() %>% 
  do.call(paste, .)

  # New York
data <- 
  neighbourhoods_table %>%  
  filter(city == "New York") %>% 
  filter(!is.na(CRI)) %>% 
  filter(active_listings >= 0) %>% 
  st_as_sf() 

streets_newyork <-
  (getbb("New York City") * c(1.01, 0.99, 0.99, 1.01)) %>%
  opq(timeout = 200) %>%
  add_osm_feature(key = "highway", value = c("primary", "secondary",
                                             "tertiary", "motorway")) %>%
  osmdata_sf()

water_NY <- 
  st_read("data/Shape_NY/NHDWaterbody.shx") %>% 
  st_as_sf() %>% 
  st_transform(102009) %>% 
  st_zm() %>% 
  st_cast("MULTIPOLYGON") %>% 
  st_make_valid() 

water_NY_ocean <- 
  st_read("data/Shape_NY/NHDArea.shp") %>% 
  st_as_sf() %>% 
  st_transform(102009) %>% 
  st_zm()

water_NY <- 
  rbind(water_NY_ocean %>% dplyr::select("geometry"),
        water_NY %>% dplyr::select("geometry"))


new_york <- 
  bivariate_mapping(data = data,
                  streets = streets_newyork,
                  water = water_NY,
                  buffer = 2000,
                  var1 = data$CSI, 
                  var2 = data$active_listings, 
                  title = "New York City",
                  xlab = "Increasing CSI", 
                  ylab = "Increasing STR Listings")


# Houston
data <- 
  neighbourhoods_table %>%  
  filter(city == "Houston") %>% 
  filter(!is.na(CRI)) %>% 
  filter(active_listings >= 0) %>% 
  st_as_sf() 

streets_houston <-
  (getbb("Houston") * c(1.01, 0.99, 0.99, 1.01)) %>%
  opq(timeout = 200) %>%
  add_osm_feature(key = "highway", value = c("primary", "secondary",
                                             "tertiary", "motorway")) %>%
  osmdata_sf()

water_TX <- 
  st_read("data/Shape_TX/NHDWaterbody.shx") %>% 
  st_as_sf() %>% 
  st_transform(102009) %>% 
  st_zm() %>% 
  st_cast("MULTIPOLYGON") %>% 
  st_make_valid() 

water_TX_ocean <- 
  st_read("data/Shape_TX/NHDArea.shp") %>% 
  st_as_sf() %>% 
  st_transform(102009) %>% 
  st_zm() %>% 
  st_make_valid()

water_TX <- 
  rbind(water_TX_ocean %>% dplyr::select("geometry"),
        water_TX %>% dplyr::select("geometry"))


houston <- 
  bivariate_mapping(data = data,
                  streets = streets_houston,
                  water = water_TX,
                  buffer = 5000,
                  var1 = data$CSI, 
                  var2 = data$active_listings, 
                  title = "Houston",
                  xlab = "Increasing CSI", 
                  ylab = "Increasing STR Listings")


  # Los Angeles
data <- 
  neighbourhoods_table %>%  
  filter(city == "Los Angeles") %>% 
  filter(!is.na(CRI)) %>% 
  filter(active_listings >= 0) %>% 
  st_as_sf()

streets_losangeles <-
  (getbb("Los Angeles") * c(1.01, 0.99, 0.99, 1.01)) %>%
  opq(memsize = 4e9) %>%
  add_osm_feature(key = "highway", value = c("primary", "secondary",
                                             "tertiary", "motorway")) %>%
  osmdata_sf()

water_CA <- 
  st_read("data/Shape_CA/NHDWaterbody.shx") %>% 
  st_as_sf() %>% 
  st_transform(102009) %>% 
  st_zm() %>% 
  st_cast("MULTIPOLYGON") %>% 
  st_make_valid() 

water_CA_ocean <- 
  st_read("data/Shape_CA/NHDArea.shp") %>% 
  st_as_sf() %>% 
  st_transform(102009) %>% 
  st_zm()

water_CA <- 
  rbind(water_CA_ocean %>% dplyr::select("geometry"),
        water_CA %>% dplyr::select("geometry"))

los_angeles <- 
  bivariate_mapping(data = data,
                  streets = streets_losangeles,
                  water = water_CA,
                  buffer = 5000,
                  var1 = data$CSI, 
                  var2 = data$active_listings, 
                  title = "Los Angeles",
                  xlab = "Increasing CSI", 
                  ylab = "Increasing STR Listings") 

# Active listings legend
bivariate_color_scale <- tibble(
  "3 - 3" = "#3F2949", # high var1, high var2
  "2 - 3" = "#435786",
  "1 - 3" = "#4885C1", # low var1, high var2
  "3 - 2" = "#77324C",
  "2 - 2" = "#806A8A", # medium var1, medium var2
  "1 - 2" = "#89A1C8",
  "3 - 1" = "#AE3A4E", # high var1, low var2
  "2 - 1" = "#BC7C8F",
  "1 - 1" = "#CABED0" # low var1, low var2
) %>%
  gather("group", "fill")

quantiles_var1 <- 
  quantile(data$CSI, probs = seq(0, 1, length.out = 4))

quantiles_var2 <- 
  quantile(data$active_listings, probs = seq(0, 1, length.out = 4))

data <- 
  data %>% 
  mutate(
    var1_quantiles = cut(
      data$CSI,
      breaks = quantiles_var1,
      include.lowest = TRUE
    ),
    var2_quantiles = cut(
      data$active_listings,
      breaks = quantiles_var2,
      include.lowest = TRUE
    ),
    group = paste(
      as.numeric(var1_quantiles), "-",
      as.numeric(var2_quantiles)
    )
  ) %>%
  left_join(bivariate_color_scale, by = "group")

bivariate_color_scale <- 
  bivariate_color_scale %>% 
  separate(group, into = c("var1", "var2"), sep = " - ") %>%
  mutate(var1 = as.integer(var1),
         var2 = as.integer(var2))

# Generate legend
legend <- 
  ggplot() +
  geom_tile(
    data = bivariate_color_scale,
    mapping = aes(
      x = var1,
      y = var2,
      fill = fill)
  ) +
  scale_fill_identity() +
  labs(x = "Increasing CSI",
       y = "Increasing active listings") +
  theme_map() +
  theme(
    axis.title = element_text(size = 10)) +
  coord_fixed()

tiff("bivariate_activelistings.tiff", units = "in", width = 15, height = 15, res = 300)

((houston/los_angeles) |
    (new_york / plot_spacer() / legend / plot_spacer() 
     + plot_layout(heights = c(3, 0.25, 0.75, 0.25)))) +
  plot_layout(ncol = 2, widths = c(1, 1)) + 
  plot_annotation(title = "The positive relationship between active listings and community sentiment\n",
                  theme = theme (text = element_text(family = "Helvetica Light",
                                                     size = 18),
                                 plot.title = element_text(hjust = 0.5)))

dev.off()

### NON MOVER
cities_non_mover <- 
  cities_table %>% 
  arrange(desc(non_mover_pct_pop)) %>% 
  dplyr::select(city) %>% 
  st_drop_geometry() %>% 
  do.call(paste, .)

  # Chicago
data <- 
  neighbourhoods_table %>%  
  filter(city == "Chicago") %>% 
  filter(!is.na(CRI)) %>% 
  filter(non_mover_pct_pop >= 0) %>% 
  st_as_sf() %>% 
  mutate(mover_pct_pop = 1/non_mover_pct_pop)

streets_chicago <-
  (getbb("Chicago") * c(1.01, 0.99, 0.99, 1.01)) %>%
  opq(timeout = 200) %>%
  add_osm_feature(key = "highway", value = c("primary", "secondary",
                                             "tertiary", "motorway")) %>%
  osmdata_sf()

water_IL <- 
  st_read("data/Shape_IL/NHDWaterbody.shx") %>% 
  st_as_sf() %>% 
  st_transform(102009) %>% 
  st_zm() %>% 
  st_cast("MULTIPOLYGON") %>% 
  st_make_valid() 

water_IL_lake <- 
  st_read("data/hydro_p_LakeMichigan/hydro_p_LakeMichigan.shp") %>% 
  st_as_sf() %>% 
  st_transform(102009) %>% 
  st_zm()

water_IL <- 
  rbind(water_IL_lake %>% dplyr::select("geometry"),
        water_IL %>% dplyr::select("geometry"))

chicago <- 
  bivariate_mapping(data = data,
                  streets = streets_chicago,
                  water = water_IL,
                  buffer = 8000,
                  var1 = data$CSI, 
                  var2 = data$mover_pct_pop, 
                  title = "Chicago",
                  xlab = "Increasing CSI", 
                  ylab = "Less Residentially Stable")


# Washington
data <- 
  neighbourhoods_table %>%  
  filter(city == "Washington") %>% 
  filter(!is.na(CRI)) %>% 
  filter(non_mover_pct_pop >= 0) %>% 
  st_as_sf() %>% 
  mutate(mover_pct_pop = 1/non_mover_pct_pop)

streets_washington <-
  (getbb("Washington DC") * c(1.01, 0.99, 0.99, 1.01)) %>%
  opq(timeout = 200) %>%
  add_osm_feature(key = "highway", value = c("primary", "secondary",
                                             "tertiary", "motorway")) %>%
  osmdata_sf()

water_DC <- 
  st_read("data/Shape_DC/NHDWaterbody.shx") %>% 
  st_as_sf() %>% 
  st_transform(102009) %>% 
  st_zm() %>% 
  st_cast("MULTIPOLYGON") %>% 
  st_make_valid() 

water_DC_ocean <- 
  st_read("data/Shape_DC/NHDArea.shp") %>% 
  st_as_sf() %>% 
  st_transform(102009) %>% 
  st_zm()

water_DC <- 
  rbind(water_DC_ocean %>% dplyr::select("geometry"),
        water_DC %>% dplyr::select("geometry"))

washington <- 
  bivariate_mapping(data = data,
                  streets = streets_washington,
                  water = water_DC,
                  buffer = 3000,
                  var1 = data$CSI, 
                  var2 = data$mover_pct_pop, 
                  title = "Washington",
                  xlab = "Increasing CSI", 
                  ylab = "Less Residentially Stable") 

  # Austin
data <- 
  neighbourhoods_table %>%  
  filter(city == "Austin") %>% 
  filter(!is.na(CRI)) %>% 
  filter(non_mover_pct_pop >= 0) %>% 
  st_as_sf() %>% 
  mutate(mover_pct_pop = 1/non_mover_pct_pop)

streets_austin <-
  (getbb("Austin") * c(1.01, 0.99, 0.99, 1.01)) %>%
  opq(timeout = 200) %>%
  add_osm_feature(key = "highway", value = c("primary", "secondary",
                                             "tertiary", "motorway")) %>%
  osmdata_sf()

water_TX <- 
  st_read("data/Shape_TX/NHDWaterbody.shx") %>% 
  st_as_sf() %>% 
  st_transform(102009) %>% 
  st_zm() %>% 
  st_cast("MULTIPOLYGON") %>% 
  st_make_valid() 

water_TX_ocean <- 
  st_read("data/Shape_TX/NHDArea.shp") %>% 
  st_as_sf() %>% 
  st_transform(102009) %>% 
  st_zm() %>% 
  st_make_valid()

water_TX_supplement <- 
  st_read("data/Surface_Water-shp/474c9f1b-1005-42f0-a925-3be3426b11c5202046-1-aqovxe.5877.shp") %>% 
  st_as_sf() %>% 
  st_transform(102009) %>% 
  st_zm() %>% 
  st_make_valid()

water_TX <- 
  rbind(water_TX_supplement %>% dplyr::select("geometry"),
        water_TX_ocean %>% dplyr::select("geometry"))

austin <- 
  bivariate_mapping(data = data,
                  streets = streets_austin,
                  water = water_TX,
                  buffer = 5000,
                  var1 = data$CSI, 
                  var2 = data$mover_pct_pop, 
                  title = "Austin",
                  xlab = "Increasing CSI", 
                  ylab = "Decreasing Residential Stability") %>% 
  plot()

# Non mover legend
bivariate_color_scale <- tibble(
  "3 - 3" = "#3F2949", # high var1, high var2
  "2 - 3" = "#435786",
  "1 - 3" = "#4885C1", # low var1, high var2
  "3 - 2" = "#77324C",
  "2 - 2" = "#806A8A", # medium var1, medium var2
  "1 - 2" = "#89A1C8",
  "3 - 1" = "#AE3A4E", # high var1, low var2
  "2 - 1" = "#BC7C8F",
  "1 - 1" = "#CABED0" # low var1, low var2
) %>%
  gather("group", "fill")

quantiles_var1 <- 
  quantile(data$CSI, probs = seq(0, 1, length.out = 4))

quantiles_var2 <- 
  quantile(data$mover_pct_pop, probs = seq(0, 1, length.out = 4))

data <- 
  data %>% 
  mutate(
    var1_quantiles = cut(
      data$CSI,
      breaks = quantiles_var1,
      include.lowest = TRUE
    ),
    var2_quantiles = cut(
      data$mover_pct_pop,
      breaks = quantiles_var2,
      include.lowest = TRUE
    ),
    group = paste(
      as.numeric(var1_quantiles), "-",
      as.numeric(var2_quantiles)
    )
  ) %>%
  left_join(bivariate_color_scale, by = "group")

bivariate_color_scale <- 
  bivariate_color_scale %>% 
  separate(group, into = c("var1", "var2"), sep = " - ") %>%
  mutate(var1 = as.integer(var1),
         var2 = as.integer(var2))

# Generate legend
legend <- 
  ggplot() +
  geom_tile(
    data = bivariate_color_scale,
    mapping = aes(
      x = var1,
      y = var2,
      fill = fill)
  ) +
  scale_fill_identity() +
  labs(x = "Increasing CSI",
       y = "Increasing residential instability") +
  theme_map() +
  theme(
    axis.title = element_text(size = 10)) +
  coord_fixed()

tiff("bivariate_nonmover.tiff", units = "in", width = 15, height = 15, res = 300)

((austin/washington) |
    (chicago / plot_spacer() / legend / plot_spacer() 
     + plot_layout(heights = c(3, 0.25, 0.75, 0.25)))) +
  plot_layout(ncol = 2, widths = c(1, 1)) + 
  plot_annotation(title = "The positive relationship between residential instability and community sentiment\n",
                  theme = theme (text = element_text(family = "Helvetica Light",
                                                     size = 18),
                                 plot.title = element_text(hjust = 0.5)))

dev.off()

# OWNER_OCCUPIED
cities_OO <- 
  cities_table %>% 
  arrange((owner_occupied_pct_household)) %>% 
  dplyr::select(city) %>% 
  st_drop_geometry() %>% 
  do.call(paste, .)

  # San Diego
data <- 
  neighbourhoods_table %>%  
  filter(city == "San Diego") %>% 
  filter(!is.na(CRI)) %>% 
  filter(owner_occupied_pct_household >= 0) %>% 
  st_as_sf() 

streets_sandiego <-
  (getbb("San Diego") * c(1.01, 0.99, 0.99, 1.01)) %>%
  opq(timeout = 200) %>%
  add_osm_feature(key = "highway", value = c("primary", "secondary",
                                             "tertiary", "motorway")) %>%
  osmdata_sf()

water_CA <- 
  st_read("data/Shape_CA/NHDWaterbody.shx") %>% 
  st_as_sf() %>% 
  st_transform(102009) %>% 
  st_zm() %>% 
  st_cast("MULTIPOLYGON") %>% 
  st_make_valid() 

water_CA_ocean <- 
  st_read("data/Shape_CA/NHDArea.shp") %>% 
  st_as_sf() %>% 
  st_transform(102009) %>% 
  st_zm() %>% 
  st_make_valid()

water_CA_ocean2 <- 
  st_read("data/ne_10m_ocean/ne_10m_ocean.shp") %>% 
  st_as_sf() %>% 
  st_transform(102009) %>% 
  st_zm() %>% 
  st_make_valid()

water_CA <- 
  rbind(water_CA_ocean %>% dplyr::select("geometry"),
        water_CA_ocean2 %>% dplyr::select("geometry"),
        water_CA %>% dplyr::select("geometry"))

san_diego <- 
        bivariate_mapping(data = data,
                  streets = streets_sandiego,
                  water = water_CA,
                  buffer = 5000,
                  var1 = data$CSI, 
                  var2 = data$owner_occupied_pct_household, 
                  title = "San Diego",
                  xlab = "Increasing CSI", 
                  ylab = "Increasing Owner-Occupied Households")

# Buffalo
data <- 
  neighbourhoods_table %>%  
  filter(city == "Buffalo") %>% 
  filter(!is.na(CRI)) %>% 
  filter(owner_occupied_pct_household >= 0) %>% 
  st_as_sf() 

streets_buffalo <-
  (getbb("Buffalo") * c(1.01, 0.99, 0.99, 1.01)) %>%
  opq(timeout = 200) %>%
  add_osm_feature(key = "highway", value = c("primary", "secondary",
                                             "tertiary", "motorway")) %>%
  osmdata_sf()

water_NY <- 
  st_read("data/Shape_NY/NHDWaterbody.shx") %>% 
  st_as_sf() %>% 
  st_transform(102009) %>% 
  st_zm() %>% 
  st_cast("MULTIPOLYGON") %>% 
  st_make_valid() 

water_NY_ocean <- 
  st_read("data/Shape_NY/NHDArea.shp") %>% 
  st_as_sf() %>% 
  st_transform(102009) %>% 
  st_zm()

water_NY <- 
  rbind(water_NY_ocean %>% dplyr::select("geometry"),
        water_NY %>% dplyr::select("geometry"))

buffalo <- 
  bivariate_mapping(data = data,
                  streets = streets_buffalo,
                  water = water_NY,
                  buffer = 5000,
                  var1 = data$CSI, 
                  var2 = data$owner_occupied_pct_household, 
                  title = "Buffalo",
                  xlab = "Increasing CSI", 
                  ylab = "Increasing Owner-Occupied Households") 

# Seattle
data <- 
  neighbourhoods_table %>%  
  filter(city == "Seattle") %>% 
  filter(!is.na(CRI)) %>% 
  filter(owner_occupied_pct_household >= 0) %>% 
  st_as_sf() 

streets_seattle <-
  (getbb("Seattle") * c(1.01, 0.99, 0.99, 1.01)) %>%
  opq(timeout = 200) %>%
  add_osm_feature(key = "highway", value = c("primary", "secondary",
                                             "tertiary", "motorway")) %>%
  osmdata_sf()

water_WA <- 
  st_read("data/Shape_WA/NHDWaterbody.shx") %>% 
  st_as_sf() %>% 
  st_transform(102009) %>% 
  st_zm() %>% 
  st_cast("MULTIPOLYGON") %>% 
  st_make_valid() 

water_WA_ocean <- 
  st_read("data/Shape_WA/NHDArea.shp") %>% 
  st_as_sf() %>% 
  st_transform(102009) %>% 
  st_zm()

water_WA <- 
  rbind(water_WA_ocean %>% dplyr::select("geometry"),
        water_WA %>% dplyr::select("geometry"))

seattle <- 
  bivariate_mapping(data = data,
                  streets = streets_seattle,
                  water = water_WA,
                  buffer = 3000,
                  var1 = data$CSI, 
                  var2 = data$owner_occupied_pct_household, 
                  title = "Seattle",
                  xlab = "Increasing CSI", 
                  ylab = "Increasing Owner Occupied Households")

# OO Legend
bivariate_color_scale <- tibble(
  "3 - 3" = "#3F2949", # high var1, high var2
  "2 - 3" = "#435786",
  "1 - 3" = "#4885C1", # low var1, high var2
  "3 - 2" = "#77324C",
  "2 - 2" = "#806A8A", # medium var1, medium var2
  "1 - 2" = "#89A1C8",
  "3 - 1" = "#AE3A4E", # high var1, low var2
  "2 - 1" = "#BC7C8F",
  "1 - 1" = "#CABED0" # low var1, low var2
) %>%
  gather("group", "fill")

quantiles_var1 <- 
  quantile(data$CSI, probs = seq(0, 1, length.out = 4))

quantiles_var2 <- 
  quantile(data$owner_occupied_pct_household, probs = seq(0, 1, length.out = 4))

data <- 
  data %>% 
  mutate(
    var1_quantiles = cut(
      data$CSI,
      breaks = quantiles_var1,
      include.lowest = TRUE
    ),
    var2_quantiles = cut(
      data$owner_occupied_pct_household,
      breaks = quantiles_var2,
      include.lowest = TRUE
    ),
    group = paste(
      as.numeric(var1_quantiles), "-",
      as.numeric(var2_quantiles)
    )
  ) %>%
  left_join(bivariate_color_scale, by = "group")

bivariate_color_scale <- 
  bivariate_color_scale %>% 
  separate(group, into = c("var1", "var2"), sep = " - ") %>%
  mutate(var1 = as.integer(var1),
         var2 = as.integer(var2))

# Generate legend
legend <- 
  ggplot() +
  geom_tile(
    data = bivariate_color_scale,
    mapping = aes(
      x = var1,
      y = var2,
      fill = fill)
  ) +
  scale_fill_identity() +
  labs(x = "Increasing CSI",
       y = "Increasing owner-occupied") +
  theme_map() +
  theme(
    axis.title = element_text(size = 10)) +
  coord_fixed()


tiff("bivariate_owneroccupied.tiff", units = "in", width = 15, height = 10, res = 300)

(((san_diego /plot_spacer() +  plot_layout(heights = c(1, 0.15))) | 
    (seattle / plot_spacer() + plot_layout(heights = c(1, 0.30)))) +
  (buffalo / plot_spacer() / legend / plot_spacer() +
     plot_layout(heights = c(3, 0.05, 0.75, 0.35)))) +
   plot_annotation(title = "The positive relationship between owner-occupied households and community sentiment\n",
                  theme = theme (text = element_text(family = "Helvetica Light",
                                                     size = 18),
                                 plot.title = element_text(hjust = 0.5)))

dev.off()

# Now, to show the absense of a relationship
# Active listings yoy 
cities_yoy <- 
  cities_table %>% 
  filter(active_listings>1000) %>% 
  arrange(desc(active_listings_yoy)) %>% 
  dplyr::select(city) %>% 
  st_drop_geometry() %>% 
  do.call(paste, .)

data <- 
  neighbourhoods_table %>%  
  filter(city == "Las Vegas") %>% 
  filter(!is.na(CRI)) %>% 
  filter(!is.na(active_listings_yoy)) %>% 
  st_as_sf() 

#quantiles_CRI <- c(-20, -2.5, 0.2, 2)

# Specify data, variables, title, labels, and quantiles (optional)
bivariate_mapping(data = data,
                  var1 = data$CRI, 
                  var2 = data$rental_pct_household, 
                  #quantiles_var1 = quantiles_CRI,
                  #quantiles_var2 = c(0, 0.00001, 0.001, 1), 
                  title = "STR Growth Rates and Community Resistance in Las Vegas",
                  xlab = "Increasing CRI", 
                  ylab = "Increasing STR Growth Rates") %>% 
  plot()

# Med income 
cities_income <- 
  cities_table %>% 
  arrange(desc(med_income)) %>% 
  dplyr::select(city) %>% 
  st_drop_geometry() %>% 
  do.call(paste, .)

data <- 
  neighbourhoods_table %>%  
  filter(city == "Detroit") %>% 
  filter(!is.na(CRI)) %>% 
  st_as_sf() 

#quantiles_CRI <- c(-20, -2.5, 0.2, 2)

# Specify data, variables, title, labels, and quantiles (optional)
bivariate_mapping(data = data,
                  var1 = data$CRI, 
                  var2 = data$med_income, 
                  #quantiles_var1 = quantiles_CRI,
                  #quantiles_var2 = c(0, 0.00001, 0.001, 1), 
                  title = "Median Income and Community Resistance in Detroit",
                  xlab = "Increasing CRI", 
                  ylab = "Increasing Median Income") %>% 
  plot()
